hds: Move v4 code to its own directory
and it just builds the C library now. The Fortran wrapper is
handled elsewhere.

dat_par.h is now a fixed file and not dynamically generated. It
is also not installed. Care must be taken to ensure that the
constants are the same in all HDS implementations.

DAT__SZLOC is only really used by the Fortran wrapper and is
a fixed size.