hds: Wrap the v4 and v5 libraries
libhds is now a wrapper library that selects the v4 or v5 API
based on the type of locator. The current implementation of the
wrapper is mostly complete but lacks support for copying between
different HDS versions and true support of cases where the API
can not tell which version to use so must try both.

hds_select.c has been modified from the automatically generated
version so do not attempt to regenerate.

HDS still has the Fortran wrapper library which links to libhds.
Most of the changes in the Fortran code involve removal of low
level HDS header files (e.g. rec.h and str.h) and switching to
use SAE for good status rathder than the (historical) private
value of DAT__OK.

HDS now depends on ONE and SAE.