figaro: Free NDF locators after each call to an action
If the application runs in monolith mode some NDF resources were
being retained. This runs in to trouble when one task opened a
file for READ and the next task in the same monolith opened the
file for UPDATE. HDF5 does not like that. Now we run ndfCancl in
the top level routine so that parameters used by the task for
NDF access will be cleared automatically.