hds: Tuning param to control V4 error reporting in threaded apps
In a multi-threaded application, the thread locking functions (datLock,
datUnlock and datLocked - which rely on V5 features) will now return
without action and without error by default if applied to a V4 locator.
If the new tuning parameter HDS-V4LOCKERROR is set non-zero, they will
instead report an error.

The thread locking functions are really only there to ensure that a
multi-threaded application is structured properly, so that an error is
reported if it attempts to modify the same HDS object from multiple
threads. Once the structure of the app has been verified using V5 HDS
files, it should be safe to process V4 files without any locking. This is
what this new tuning parameter allows.